Initial commit
This contains the initial html pages for the Sitter Sign_Up and the Sitter Search I also have the .js file for the sitter for logging the sitter sign-up data to the api. I have plenty of work to do but will have tested files ready for class on Tuesday.